+++ agent/lib/kontena-agent.rb
@@ -17,6 +17,7 @@ require_relative 'kontena/websocket_client'
 
 require_relative 'kontena/network_adapters/weave'
 require_relative 'kontena/network_adapters/ipam_client'
+require_relative 'kontena/network_adapters/ipam_cleaner'
 
 require_relative 'kontena/launchers/etcd'
 require_relative 'kontena/launchers/cadvisor'
